# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: GradeEscolar — Sistema Completo de Gestão de Horários Escolares
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /turmas/{id}/horario:
    get:
      operationId: getTurmaHorario
      tags: [turmas]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Grade horária
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HorarioSlot"

  # ── HORÁRIOS ──────────────────────────────────────────────────────────────────
  /horarios:
    get:
      operationId: listHorarios
      tags: [horarios]
      parameters:
        - name: turmaId
          in: query
          schema:
            type: integer
        - name: professorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HorarioSlot"
    post:
      operationId: createHorario
      tags: [horarios]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HorarioSlotInput"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HorarioSlot"
  /horarios/{id}:
    delete:
      operationId: deleteHorario
      tags: [horarios]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removido
  /horarios/gerar:
    post:
      operationId: gerarHorario
      tags: [horarios]
      summary: Gera horário automaticamente com algoritmo avançado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GerarHorarioInput"
      responses:
        "200":
          description: Resultado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GerarHorarioResult"
  /horarios/experimentais:
    get:
      operationId: listHorariosExperimentais
      tags: [horarios]
      summary: Lista horários experimentais
      parameters:
        - name: turmaId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HorarioExperimental"
    post:
      operationId: createHorarioExperimental
      tags: [horarios]
      summary: Cria horário experimental
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HorarioExperimentalInput"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HorarioExperimental"
  /horarios/experimentais/{nome}/promover:
    post:
      operationId: promoverHorarioExperimental
      tags: [horarios]
      summary: Promove horário experimental para oficial
      parameters:
        - name: nome
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Promovido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GerarHorarioResult"
  /horarios/experimentais/{nome}:
    delete:
      operationId: deleteHorarioExperimental
      tags: [horarios]
      summary: Remove horário experimental
      parameters:
        - name: nome
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removido

  # ── SALAS ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Professor:
      type: object
      required: [id, nome, email, ativo, createdAt]
      properties:
        id: { type: integer }
        nome: { type: string }
        email: { type: string }
        telefone: { type: ["string", "null"] }
        ativo: { type: boolean }
        disciplinaIds:
          type: array
          items: { type: integer }
        createdAt: { type: string }
...
    Disciplina:
      type: object
      required: [id, nome, cargaSemanal, cor, createdAt]
      properties:
        id: { type: integer }
        nome: { type: string }
        cargaSemanal: { type: integer }
        cor: { type: string }
        createdAt: { type: string }
...
    Turma:
      type: object
      required: [id, nome, serie, turno, anoLetivo, createdAt]
      properties:
        id: { type: integer }
        nome: { type: string }
        serie: { type: string }
        turno: { type: string, enum: [matutino, vespertino, noturno, integral] }
        anoLetivo: { type: integer }
        modalidade: { type: ["string", "null"] }
        matrizCurricularId: { type: ["integer", "null"], description: "Matriz Curricular aplicada a esta turma (RF-TUR-02), se houver." }
        disciplinaIds:
          type: array
          items: { type: integer }
        disciplinasComCarga:
          type: array
          description: "Carga horária semanal efetiva por disciplina (override da matriz aplicada, com fallback para a carga global)."
          items:
            type: object
            required: [disciplinaId, nome, cargaHorariaSemanal, origemMatriz]
            properties:
              disciplinaId: { type: integer }
              nome: { type: string }
              cargaHorariaSemanal: { type: integer }
              origemMatriz: { type: boolean }
        createdAt: { type: string }
...
    HorarioSlot:
      type: object
      required: [id, turmaId, disciplinaId, professorId, diaSemana, numeroAula]
      properties:
        id: { type: integer }
        turmaId: { type: integer }
        turma:
          $ref: "#/components/schemas/Turma"
        disciplinaId: { type: integer }
        disciplina:
          $ref: "#/components/schemas/Disciplina"
        professorId: { type: integer }
        professor:
          $ref: "#/components/schemas/Professor"
        diaSemana: { type: integer, minimum: 0, maximum: 4 }
        numeroAula: { type: integer, minimum: 1, maximum: 8 }
        sala: { type: ["string", "null"] }
    HorarioSlotInput:
      type: object
      required: [turmaId, disciplinaId, professorId, diaSemana, numeroAula]
      properties:
        turmaId: { type: integer }
        disciplinaId: { type: integer }
        professorId: { type: integer }
        diaSemana: { type: integer, minimum: 0, maximum: 4 }
        numeroAula: { type: integer, minimum: 1, maximum: 8 }
        sala: { type: string }
    HorarioExperimental:
      type: object
      required: [id, nome, turmaId, disciplinaId, professorId, diaSemana, numeroAula]
      properties:
        id: { type: integer }
        nome: { type: string }
        turmaId: { type: integer }
        disciplinaId: { type: integer }
        professorId: { type: integer }
        diaSemana: { type: integer }
        numeroAula: { type: integer }
        sala: { type: ["string", "null"] }
        ativo: { type: boolean }
    HorarioExperimentalInput:
      type: object
      required: [nome, turmaId, disciplinaId, professorId, diaSemana, numeroAula]
      properties:
        nome: { type: string }
        turmaId: { type: integer }
        disciplinaId: { type: integer }
        professorId: { type: integer }
        diaSemana: { type: integer }
        numeroAula: { type: integer }
        sala: { type: string }
    GerarHorarioInput:
      type: object
      required: [turmaId]
      properties:
        turmaId: { type: integer }
        aulaspordia: { type: integer, minimum: 4, maximum: 8 }
        substituir: { type: boolean }
        reduzirJanelas: { type: boolean }
        fatorPedagogico: { type: boolean }
        experimental: { type: boolean }
        nomeExperimental: { type: string }
    GerarHorarioResult:
      type: object
      required: [slotsGerados, conflitos]
      properties:
        slotsGerados: { type: integer }
        conflitos:
          type: array
          items: { type: string }
        horario:
          type: array
          items:
            $ref: "#/components/schemas/HorarioSlot"
...